Revert "Refactor of the buttons, code logic and comments"
